Renderer: replace array of uniforms with a UBO for the command buffer
Previously the command data (array of vec4) was hardcoded as an array
of uniform inputs in the shader. This was temporary because it is
limited to gl.MAX_FRAGMENT_UNIFORM_VECTORS which is only guaranteed
to be 224 and practically can be only 1024. Given that shapes have 3-4
commands in average, that's not enough.

Now the command buffer is backed by a Uniform Buffer Object (lacking
Buffer Textures or 1D image samplers) which provides strong typing for
the shader. Size is set to 4096 commands, which is 4x the guaranteed
gl.MAX_UNIFORM_BLOCK_SIZE.