Renderer: temporary fix for slow loading of images
There is a slowness likely caused by redrawing an array texture
at the same time that its slices are being filled with data, or
by a big number of concurrent draw requests in a small amount of
time. To avoid this problem, an image bundle only requests a
redraw when all its images are done loading.

TODO: a better way would be to batch the data transfers and the
draw requests so that there is some progress feedback for users.