Handle fetching or parsing errors for casting
Casting urls are fetched in a loop based on all sequences of a
project. If a sequence does not have shots with cast assets, the url
won't actually exist, causing a bad response (404 in a production
server, or a plain html response in development).
For the time being we simply acknowledge the response.